+++ src/doc_kg/__init__.py
@@ -2,6 +2,6 @@
 
 from doc_kg.kg import DocKG
 
-__version__ = "0.9.1"
+__version__ = "0.9.2"
 
 __all__ = ["DocKG", "__version__"]